Consolidate home page routing to fix build output
